requirements.txt
altgraph==0.17.3
Pillow==10.0.0
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.5
